Enabled users to add items. Enabled users to delete and modify an item they own

+++ Services/ItemService.cs
@@ -1,17 +1,20 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using SwapApp.Authorization;
 using SwapApp.Entities;
 using SwapApp.Exceptions;
 using SwapApp.Models;
+using System.Security.Claims;
 
 namespace SwapApp.Services
 {
     public interface IItemService
     {
-        int AddItem(AddItemDto addItem);
+        int AddItem(AddItemDto addItem, int userId);
         IEnumerable<GetItemDto> GetAllItems();
         GetItemDto GetItemById(int id);
-        bool DeleteItem (int id);
-        bool UpdateItem(UpdateItemDto updateItem, int id);
+        bool DeleteItem (int id, ClaimsPrincipal user);
+        bool UpdateItem(UpdateItemDto updateItem, int id, ClaimsPrincipal user);
 
     }
 
@@ -20,19 +23,28 @@ namespace SwapApp.Services
         private readonly ItemDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<ItemService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ItemService(ItemDbContext dbContext, IMapper mapper, ILogger<ItemService> logger)
+        public ItemService(ItemDbContext dbContext, IMapper mapper, ILogger<ItemService> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public bool UpdateItem (UpdateItemDto updateItem, int id)
+        public bool UpdateItem (UpdateItemDto updateItem, int id, ClaimsPrincipal user)
         {
             var item = _dbContext.Item.FirstOrDefault(x => x.Id == id);
             if (item is null)
                 throw new NotFoundException("Item not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, item, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You can't update this item.");
+            }
 
             item.Name = updateItem.Name;
             item.Description = updateItem.Description;
@@ -46,11 +58,19 @@ namespace SwapApp.Services
             return true;
         }
 
-        public bool DeleteItem (int id)
+        public bool DeleteItem (int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Item with id: {id} DELETE action invoked");
             var item = _dbContext.Item.FirstOrDefault(x => x.Id == id);
             if (item is null) return false;
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, item, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You can't delete this item");
+            }
+
             _dbContext.Item.Remove(item);
             _dbContext.SaveChanges();
             return true;
@@ -74,9 +94,10 @@ namespace SwapApp.Services
 
         }
 
-        public int AddItem(AddItemDto addItem)
+        public int AddItem(AddItemDto addItem, int userId)
         {
             var item = _mapper.Map<Item>(addItem);
+            item.UserId = userId;
             _dbContext.Item.Add(item);
             _dbContext.SaveChanges();
             return item.Id;
